refactor(web): move dismissible callout to design system
